Initial project structure and API Specification
This commit is an initial project structure and the Tern REST API
Specification proposal.

This creates the initial API Specification for the Tern, focused on
the Tern reports (as JSON).
The implementation is not done in this commit, as before
development/implementing the code, we want to agree on the API
Specification first.
The image report data model is incomplete
(``api.v1.commom_models.image_report_data``), and it is intentional as
the goal is to validate the high-level API design.

This commit also provides a basic tests setup to create the standards
to test using the API as an entry-point.
The CI for basic tests, linting, and offline documentation generation.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>