Implement the report using the Tern
This commit implements the report using Tern
The reports use the ``concurrent.futures``, but instead of building
all the implementation for it, we use flask-executer that makes the
implementation on top of Flask.

All the report requests go to the ``tern-tasks`` (Futures) and are
handled such as a Celery implementation.
https://docs.celeryproject.org/en/stable/reference/celery.states.html#misc
All tasks start with the status 'PENDING', which means that the task
has no defined status until the ``tern-tasks`` knows it.
Once the task is in picked up by the ``tern-tasks`` it is moved to
'RUNNING'.
If the processing finishes, the status is moved to 'SUCCESS'
independently of the task content/result. Here the status is for
the task itself.
The 'FAILURE' status is given when the task cannot finish its
routine.

The Tern, for now, is called using ``subprocess``, and the output is
handled by the stdout and stderr.
The initial API specification was changed a bit for more clearness.

How to run as Docker container was added to the ``README.md``.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>